Updated for loop to increment instead of for .. in
There is potential for a bug when using for .. in loops over array like things. 

http://stackoverflow.com/questions/500504/why-is-using-for-in-with-array-iteration-such-a-bad-idea

This became a reality for one user who has extended the Array prototype. 

http://stackoverflow.com/questions/23353013/npm-install-authy-api-typeerror-in-nodejs

This will fix the bug that the user is experiencing and prevent other changes to the Array prototype from breaking the loop.